# TST: Speedup tests