--- a/tests/map/test_map.py
+++ b/tests/map/test_map.py
@@ -955,7 +955,7 @@ def test_add_mapspec_axis_from_step(tmp_path: Path) -> None:
         tmp_path,
         internal_shapes=internal_shapes_map,  # type: ignore[arg-type]
         parallel=False,
-        storage=storage,
+        storage="dict",
     )
     assert results["sum"].output.tolist() == [13]
     load_xarray_dataset(run_folder=tmp_path)
@@ -1683,7 +1683,7 @@ def test_pipeline_with_heterogeneous_chunksize(chunksizes):
 
     pipeline = Pipeline([f, g, h])
     inputs = {"x": [1, 2, 3]}
-    results = pipeline.map(inputs, chunksizes=chunksizes)
+    results = pipeline.map(inputs, chunksizes=chunksizes, parallel=False, storage="dict")
     assert results["y1"].output.tolist() == [0, 1, 2]
     assert results["z"].output.tolist() == [2, 3, 4]
     assert results["y2"].output.tolist() == [2, 3, 4]
@@ -1803,11 +1803,11 @@ def test_profiling_and_parallel_unsupported_warning() -> None:
 
     test_pipeline_with_profile_true = Pipeline([a], profile=True)
     with pytest.warns(UserWarning, match="`profile=True` is not supported with `parallel=True`"):
-        test_pipeline_with_profile_true.map({"val": np.array([1, 2, 3])})
+        test_pipeline_with_profile_true.map({"val": [1]})
 
     test_pipeline_with_profile_none = Pipeline([a_profile], profile=None)
     with pytest.warns(UserWarning, match="`profile=True` is not supported with `parallel=True`"):
-        test_pipeline_with_profile_none.map({"val": np.array([1, 2, 3])})
+        test_pipeline_with_profile_none.map({"val": [1]})
 
 
 def test_map_with_auto_subpipeline(tmp_path: Path):