--- a/tests/map/test_run_dynamic_internal_shape.py
+++ b/tests/map/test_run_dynamic_internal_shape.py
@@ -520,7 +520,7 @@ def test_simple_2d(
 
 @pytest.mark.parametrize("return_results", [True, False])
 @pytest.mark.parametrize("parallel", [True, False])
-@pytest.mark.parametrize("storage", ["file_array", "dict", "shared_memory_dict"])
+@pytest.mark.parametrize("storage", ["file_array", "dict"])
 def test_multiple_outputs_with_dynamic_shape_and_individual_outputs_are_nd_arrays(
     tmp_path: Path,
     return_results: bool,  # noqa: FBT001