--- a/tests/test_pipeline.py
+++ b/tests/test_pipeline.py
@@ -748,7 +748,7 @@ def test_parameterless_pipefunc() -> None:
     assert pipeline() == 1
     assert pipeline.topological_generations.root_args == []
     assert pipeline.topological_generations.function_lists == [[pipeline["c"]]]
-    r = pipeline.map({})
+    r = pipeline.map({}, storage="dict", parallel=False)
     assert r["c"].output == 1
 
     @pipefunc(output_name="d")
@@ -767,7 +767,7 @@ def test_parameterless_pipefunc() -> None:
         [pipeline["c"], pipeline["d"]],
         [pipeline["e"]],
     ]
-    r = pipeline.map({})
+    r = pipeline.map({}, storage="dict", parallel=False)
     assert r["e"].output == 3
 
 
@@ -1008,7 +1008,7 @@ def test_nested_pipefunc_in_pipeline_renames() -> None:
     assert func.pipeline.functions[0].renames == {}
     r = pipeline.run("test.y", kwargs={"test.n": 2})
     assert r == 8
-    r = pipeline.map(inputs={"test.n": 2})
+    r = pipeline.map(inputs={"test.n": 2}, storage="dict", parallel=False)
     assert r["test.y"].output == 8