Currently there are a lot of slow tests:
```bash
33.87s call     tests/map/test_adaptive.py::test_adaptive_run_dynamic_internal_shape
15.87s call     tests/map/test_map.py::test_add_mapspec_axis_from_step[shared_memory_dict]
11.56s call     tests/map/test_map.py::test_add_mapspec_axis_from_step[zarr_shared_memory]
11.31s call     tests/map/test_run_dynamic_internal_shape.py::test_multiple_outputs_with_dynamic_shape_and_individual_outputs_are_nd_arrays[shared_memory_dict-True-True]
8.68s call     tests/test_pipeline.py::test_pipeline_map_zero_size
8.05s call     tests/map/test_run_dynamic_internal_shape.py::test_multiple_outputs_with_dynamic_shape_and_individual_outputs_are_nd_arrays[shared_memory_dict-True-False]
6.87s call     tests/map/test_map.py::test_profiling_and_parallel_unsupported_warning
6.83s call     tests/map/test_map.py::test_parallel_memory_storage[zarr_memory]
6.13s call     tests/map/test_map.py::test_parallel_memory_storage[dict]
4.99s call     tests/map/test_map.py::test_pipeline_with_heterogeneous_executor
4.87s call     tests/map/test_map.py::test_pipeline_with_heterogeneous_chunksize[chunksizes0]
4.56s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_fixed_indices
4.25s call     tests/map/test_map.py::test_map_async_with_progress[eager]
4.23s call     tests/map/test_map.py::test_map_async_with_progress[generation]
4.21s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_complex_mapspec
4.13s call     tests/map/test_run_dynamic_internal_shape.py::test_dynamic_internal_shape_async[None-False]
4.06s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_without_returning_results
4.05s call     tests/test_pipeline.py::test_double_output_then_iterate_over_single_axis
4.00s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_internal_shapes
4.00s call     tests/map/test_run_dynamic_internal_shape.py::test_dynamic_internal_shape_async[?-False]
3.95s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_persist_memory[True]
3.91s call     tests/map/test_map.py::test_pipeline_with_heterogeneous_chunksize[1]
3.78s call     tests/test_pipeline_mapspec.py::test_mapping_over_default
3.75s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_mapspec
3.75s call     tests/map/test_run_dynamic_internal_shape.py::test_dynamic_internal_shape_async[None-True]
3.75s call     tests/map/test_run_dynamic_internal_shape.py::test_dynamic_internal_shape_async[?-True]
3.66s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_different_storage[shared_memory_dict]
3.64s call     tests/test_pipeline.py::test_parameterless_pipefunc
3.62s call     tests/map/test_eager_scheduler.py::test_complex_dependency_graph
3.41s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_custom_executor
3.29s call     tests/test_pipeline_mapspec.py::test_calling_add_with_autogen_mapspec[3]
3.10s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_different_storage[dict]
3.05s call     tests/map/test_run_dynamic_internal_shape.py::test_multiple_outputs_with_dynamic_shape_and_individual_outputs_are_nd_arrays[shared_memory_dict-False-True]
2.85s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_persist_memory[False]
2.80s call     tests/map/test_map.py::test_parallel
2.74s call     tests/test_cache.py::test_pickling_and_deepcopy_lru[True]
2.74s call     tests/test_plotting.py::test_plot_with_defaults[holoviews]
2.71s call     tests/test_pipeline_update.py::test_scope_with_mapspec
2.66s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_diamond_pattern
2.48s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_chunksizes
```

After 
```bash
6.50s call     tests/map/test_map.py::test_parallel_memory_storage[zarr_memory]
5.98s call     tests/map/test_map.py::test_profiling_and_parallel_unsupported_warning
5.55s call     tests/map/test_map.py::test_parallel_memory_storage[dict]
4.56s call     tests/map/test_map.py::test_parallel
3.37s call     tests/test_pipeline_mapspec.py::test_mapping_over_default
3.14s call     tests/map/test_run_dynamic_internal_shape.py::test_multiple_outputs_with_dynamic_shape_and_individual_outputs_are_nd_arrays[dict-True-False]
3.11s call     tests/map/test_run_dynamic_internal_shape.py::test_multiple_outputs_with_dynamic_shape_and_individual_outputs_are_nd_arrays[dict-True-True]
2.97s call     tests/test_pipeline_mapspec.py::test_with_progress
2.78s call     tests/map/test_eager_scheduler.py::test_eager_scheduler_with_diamond_pattern
2.66s call     tests/map/test_map.py::test_storage_options_zarr_memory_parallel
2.52s call     tests/map/test_run_dynamic_internal_shape.py::test_multiple_outputs_with_dynamic_shape_and_individual_outputs_are_nd_arrays[file_array-True-False]
2.51s call     tests/map/test_map.py::test_pipeline_with_heterogeneous_chunksize[1]
2.45s setup    tests/test_variant_pipeline_widgets.py::test_visualize_default_backend
2.42s call     tests/map/test_map.py::test_map_with_partial
2.39s setup    tests/test_variant_pipeline_widgets.py::test_visualize_with_variant_selection
2.37s setup    tests/test_variant_pipeline_widgets.py::test_repr_mimebundle_without_ipywidgets_or_rich
```